fix: widen scroll/scroll_offset/prev_total_lines from u16 to u32
Root cause: scroll_offset was u16, wrapped to 0 after ~65535 evictions,
breaking PageUp/PageDown. Now all scroll coordinates are u32:
- scroll: Cell<u32>, prev_total_lines: Cell<u32>
- RenderState::scroll_offset: u32, prev_total_lines: u32
- ratatui boundary clamps to u16 via min(u16::MAX) as u16

+++ src/ui/render_state.rs
@@ -29,10 +29,10 @@ pub struct RenderState {
     pub menu_items: Vec<(String, String)>,
     pub menu_idx: usize,
     pub menu_title: Option<String>,
-    pub scroll_offset: u16,
+    pub scroll_offset: u32,
     pub follow_tail: bool,
     #[allow(dead_code)]
-    pub prev_total_lines: u16,
+    pub prev_total_lines: u32,
     pub unseen_lines: u32,
     pub evicted_lines: u64,
     pub buffer_total_lines: u64,
